Change RGBA field to use uint32

+++ zivid_camera/src/zivid_camera.cpp
@@ -737,7 +737,7 @@ void ZividCamera::publishPointCloudXYZRGBA(
   msg->fields.push_back(createPointField("x", 0, sensor_msgs::msg::PointField::FLOAT32, 1));
   msg->fields.push_back(createPointField("y", 4, sensor_msgs::msg::PointField::FLOAT32, 1));
   msg->fields.push_back(createPointField("z", 8, sensor_msgs::msg::PointField::FLOAT32, 1));
-  msg->fields.push_back(createPointField("rgba", 12, sensor_msgs::msg::PointField::FLOAT32, 1));
+  msg->fields.push_back(createPointField("rgba", 12, sensor_msgs::msg::PointField::UINT32, 1));
   msg->is_dense = false;
 
   // Note that the "rgba" field is actually byte order "bgra" on little-endian systems. For this
